refactor(폴더구조): 재사용성 없는 컴포넌트들을 page폴더로 이동
