+++ src/pages/CardRegisterForm/CardNumberInputField.tsx
@@ -1,21 +1,15 @@
 import React from "react";
-import InputField from "../components/InputField/InputField";
+import InputField from "../../components/InputField/InputField";
 import useInputField, { IndividualValidator } from "@/hooks/useInputField";
 import useValidation from "@/hooks/useValidation";
 import useInput from "@/hooks/useInput";
 
-const VALID_LENGTH = 2;
-const INPUTS_COUNT = 2;
+const VALID_LENGTH = 4;
+const INPUTS_COUNT = 4;
 const individualValidators: IndividualValidator[] = [
-  {
-    errorMessage: "월은 1~12의 범위여야합니다.",
-    validate: (month: string) => month === "" || (Number(month) >= 1 && Number(month) <= 12),
-    index: [0],
-  },
   {
     errorMessage: `길이는 ${VALID_LENGTH}여야합니다.`,
     validate: (input: string) => input.length === 0 || input.length === VALID_LENGTH,
-    index: [0, 1],
   },
   {
     errorMessage: `입력은 숫자형이어야합니다.`,
@@ -23,11 +17,12 @@ const individualValidators: IndividualValidator[] = [
   },
 ];
 
-const ExpirationDateInputField = ({ reduceds }: { reduceds: ReturnType<typeof useInput>[] }) => {
+const CardNumberInputFieldCopy = ({ reduceds }: { reduceds: ReturnType<typeof useInput>[] }) => {
   const validationStates = reduceds.map((reduced) => useValidation(reduced, individualValidators));
+
   return (
     <InputField>
-      <InputField.Label>유효기간</InputField.Label>
+      <InputField.Label>카드번호</InputField.Label>
       <InputField.Inputs>
         {Array.from({ length: INPUTS_COUNT }).map((_, index) => (
           <InputField.Input
@@ -48,4 +43,4 @@ const ExpirationDateInputField = ({ reduceds }: { reduceds: ReturnType<typeof us
   );
 };
 
-export default ExpirationDateInputField;
+export default CardNumberInputFieldCopy;
