+++ src/pages/CardRegisterForm/ExpirationDateInputField.tsx
@@ -1,27 +1,33 @@
 import React from "react";
-import InputField from "../components/InputField/InputField";
-import { IndividualValidator } from "@/hooks/useInputField";
-import useInput from "@/hooks/useInput";
+import InputField from "../../components/InputField/InputField";
+import useInputField, { IndividualValidator } from "@/hooks/useInputField";
 import useValidation from "@/hooks/useValidation";
+import useInput from "@/hooks/useInput";
 
-const VALID_LENGTH = 30;
-const INPUTS_COUNT = 1;
+const VALID_LENGTH = 2;
+const INPUTS_COUNT = 2;
 const individualValidators: IndividualValidator[] = [
   {
-    validate: (input: string) => /^[a-zA-Z\s]*$/.test(input),
-    errorMessage: "이름은 영어 대문자로 입력해주세요.",
+    errorMessage: "월은 1~12의 범위여야합니다.",
+    validate: (month: string) => month === "" || (Number(month) >= 1 && Number(month) <= 12),
+    index: [0],
   },
   {
-    validate: (input: string) => !/\s{2,}/.test(input),
-    errorMessage: "이름의 공백은 2회이상 연속되지 않아야 합니다..",
+    errorMessage: `길이는 ${VALID_LENGTH}여야합니다.`,
+    validate: (input: string) => input.length === 0 || input.length === VALID_LENGTH,
+    index: [0, 1],
+  },
+  {
+    errorMessage: `입력은 숫자형이어야합니다.`,
+    validate: (input: string) => input.length === 0 || /^[0-9]*$/.test(input),
   },
 ];
 
-const OwnerNameInputField = ({ reduceds }: { reduceds: ReturnType<typeof useInput>[] }) => {
+const ExpirationDateInputField = ({ reduceds }: { reduceds: ReturnType<typeof useInput>[] }) => {
   const validationStates = reduceds.map((reduced) => useValidation(reduced, individualValidators));
   return (
     <InputField>
-      <InputField.Label>소유자 이름</InputField.Label>
+      <InputField.Label>유효기간</InputField.Label>
       <InputField.Inputs>
         {Array.from({ length: INPUTS_COUNT }).map((_, index) => (
           <InputField.Input
@@ -29,7 +35,6 @@ const OwnerNameInputField = ({ reduceds }: { reduceds: ReturnType<typeof useInpu
             isError={!validationStates[index].inputState.isValid}
             onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
               if (event.target.value.length > VALID_LENGTH) return;
-              event.target.value = event.target.value.toUpperCase();
               validationStates[index].setValue(event.target.value);
             }}
             value={validationStates[index].inputState.value}
@@ -43,4 +48,4 @@ const OwnerNameInputField = ({ reduceds }: { reduceds: ReturnType<typeof useInpu
   );
 };
 
-export default OwnerNameInputField;
+export default ExpirationDateInputField;
